requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.5.2
mccabe==0.6.1
multidict==5.1.0
requests==2.25.1
six==1.15.0
snowballstemmer==2.0.0
tabulate==0.8.7
toml==0.10.2
urllib3==1.26.2
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
arrow==0.17.0